// packages/services/shopify/src/__generated__/graphql.ts
export type Collection = HasMetafields & Node & OnlineStorePublishable & Trackable & {
  readonly __typename: 'Collection';
  /** Stripped description of the collection, single line with HTML tags removed. */
  readonly description: Scalars['String']['output'];
  /** The description of the collection, complete with HTML formatting. */
  readonly descriptionHtml: Scalars['HTML']['output'];
  /**
   * A human-friendly unique string for the collection automatically generated from its title.
   * Limit of 255 characters.
   *
   */
  readonly handle: Scalars['String']['output'];
  /** A globally-unique ID. */
  readonly id: Scalars['ID']['output'];
  /** Image associated with the collection. */
  readonly image: Maybe<Image>;
  /** A [custom field](https://shopify.dev/docs/apps/build/custom-data), including its `namespace` and `key`, that's associated with a Shopify resource for the purposes of adding and storing additional information. */
  readonly metafield: Maybe<Metafield>;
  /** A list of [custom fields](/docs/apps/build/custom-data) that a merchant associates with a Shopify resource. */
  readonly metafields: ReadonlyArray<Maybe<Metafield>>;
  /** The URL used for viewing the resource on the shop's Online Store. Returns `null` if the resource is currently not published to the Online Store sales channel. */
  readonly onlineStoreUrl: Maybe<Scalars['URL']['output']>;
  /** List of products in the collection. */
  readonly products: ProductConnection;
  /** The collection's SEO information. */
  readonly seo: Seo;
  /** The collection’s name. Limit of 255 characters. */
  readonly title: Scalars['String']['output'];
  /** URL parameters to be added to a page URL to track the origin of on-site search traffic for [analytics reporting](https://help.shopify.com/manual/reports-and-analytics/shopify-reports/report-types/default-reports/behaviour-reports). Returns a result when accessed through the [search](https://shopify.dev/docs/api/storefront/current/queries/search) or [predictiveSearch](https://shopify.dev/docs/api/storefront/current/queries/predictiveSearch) queries, otherwise returns null. */
  readonly trackingParameters: Maybe<Scalars['String']['output']>;
  /** The date and time when the collection was last modified. */
  readonly updatedAt: Scalars['DateTime']['output'];
};
// ...
export type CollectionConnection = {
  readonly __typename: 'CollectionConnection';
  /** A list of edges. */
  readonly edges: ReadonlyArray<CollectionEdge>;
  /** A list of the nodes contained in CollectionEdge. */
  readonly nodes: ReadonlyArray<Collection>;
  /** Information to aid in pagination. */
  readonly pageInfo: PageInfo;
  /** The total count of Collections. */
  readonly totalCount: Scalars['UnsignedInt64']['output'];
};

/**
 * An auto-generated type which holds one Collection and a cursor during pagination.
 *
 */
export type CollectionEdge = {
  readonly __typename: 'CollectionEdge';
  /** A cursor for use in pagination. */
  readonly cursor: Scalars['String']['output'];
  /** The item at the end of CollectionEdge. */
  readonly node: Collection;
};
// ...
export type GenericFile = Node & {
  readonly __typename: 'GenericFile';
  /** A word or phrase to indicate the contents of a file. */
  readonly alt: Maybe<Scalars['String']['output']>;
  /** A globally-unique ID. */
  readonly id: Scalars['ID']['output'];
  /** The MIME type of the file. */
  readonly mimeType: Maybe<Scalars['String']['output']>;
  /** The size of the original file in bytes. */
  readonly originalFileSize: Maybe<Scalars['Int']['output']>;
  /** The preview image for the file. */
  readonly previewImage: Maybe<Image>;
  /** The URL of the file. */
  readonly url: Maybe<Scalars['URL']['output']>;
};
// ...
export type Image = {
  readonly __typename: 'Image';
  /** A word or phrase to share the nature or contents of an image. */
  readonly altText: Maybe<Scalars['String']['output']>;
  /** The original height of the image in pixels. Returns `null` if the image isn't hosted by Shopify. */
  readonly height: Maybe<Scalars['Int']['output']>;
  /** A unique ID for the image. */
  readonly id: Maybe<Scalars['ID']['output']>;
  /**
   * The location of the original image as a URL.
   *
   * If there are any existing transformations in the original source URL, they will remain and not be stripped.
   *
   * @deprecated Use `url` instead.
   */
  readonly originalSrc: Scalars['URL']['output'];
  /**
   * The location of the image as a URL.
   * @deprecated Use `url` instead.
   */
  readonly src: Scalars['URL']['output'];
  /**
   * The ThumbHash of the image.
   *
   * Useful to display placeholder images while the original image is loading.
   *
   * See https://evanw.github.io/thumbhash/ for details on how to use it.
   *
   */
  readonly thumbhash: Maybe<Scalars['String']['output']>;
  /**
   * The location of the transformed image as a URL.
   *
   * All transformation arguments are considered "best-effort". If they can be applied to an image, they will be.
   * Otherwise any transformations which an image type doesn't support will be ignored.
   *
   * @deprecated Use `url(transform:)` instead
   */
  readonly transformedSrc: Scalars['URL']['output'];
  /**
   * The location of the image as a URL.
   *
   * If no transform options are specified, then the original image will be preserved including any pre-applied transforms.
   *
   * All transformation options are considered "best-effort". Any transformation that the original image type doesn't support will be ignored.
   *
   * If you need multiple variations of the same image, then you can use [GraphQL aliases](https://graphql.org/learn/queries/#aliases).
   *
   */
  readonly url: Scalars['URL']['output'];
  /** The original width of the image in pixels. Returns `null` if the image isn't hosted by Shopify. */
  readonly width: Maybe<Scalars['Int']['output']>;
};
// ...
export type MediaImage = Media & Node & {
  readonly __typename: 'MediaImage';
  /** A word or phrase to share the nature or contents of a media. */
  readonly alt: Maybe<Scalars['String']['output']>;
  /** A globally-unique ID. */
  readonly id: Scalars['ID']['output'];
  /** The image for the media. */
  readonly image: Maybe<Image>;
  /** The media content type. */
  readonly mediaContentType: MediaContentType;
  /** The presentation for a media. */
  readonly presentation: Maybe<MediaPresentation>;
  /** The preview image for the media. */
  readonly previewImage: Maybe<Image>;
};
// ...
export type Metafield = Node & {
  readonly __typename: 'Metafield';
  /** The date and time when the storefront metafield was created. */
  readonly createdAt: Scalars['DateTime']['output'];
  /** The description of a metafield. */
  readonly description: Maybe<Scalars['String']['output']>;
  /** A globally-unique ID. */
  readonly id: Scalars['ID']['output'];
  /** The unique identifier for the metafield within its namespace. */
  readonly key: Scalars['String']['output'];
  /** The container for a group of metafields that the metafield is associated with. */
  readonly namespace: Scalars['String']['output'];
  /** The type of resource that the metafield is attached to. */
  readonly parentResource: MetafieldParentResource;
  /** Returns a reference object if the metafield's type is a resource reference. */
  readonly reference: Maybe<MetafieldReference>;
  /** A list of reference objects if the metafield's type is a resource reference list. */
  readonly references: Maybe<MetafieldReferenceConnection>;
  /**
   * The type name of the metafield.
   * Refer to the list of [supported types](https://shopify.dev/apps/metafields/definitions/types).
   *
   */
  readonly type: Scalars['String']['output'];
  /** The date and time when the metafield was last updated. */
  readonly updatedAt: Scalars['DateTime']['output'];
  /** The data stored in the metafield. Always stored as a string, regardless of the metafield's type. */
  readonly value: Scalars['String']['output'];
};
// ...
export type Metaobject = Node & OnlineStorePublishable & {
  readonly __typename: 'Metaobject';
  /** Accesses a field of the object by key. */
  readonly field: Maybe<MetaobjectField>;
  /**
   * All object fields with defined values.
   * Omitted object keys can be assumed null, and no guarantees are made about field order.
   *
   */
  readonly fields: ReadonlyArray<MetaobjectField>;
  /** The unique handle of the metaobject. Useful as a custom ID. */
  readonly handle: Scalars['String']['output'];
  /** A globally-unique ID. */
  readonly id: Scalars['ID']['output'];
  /** The URL used for viewing the metaobject on the shop's Online Store. Returns `null` if the metaobject definition doesn't have the `online_store` capability. */
  readonly onlineStoreUrl: Maybe<Scalars['URL']['output']>;
  /**
   * The metaobject's SEO information. Returns `null` if the metaobject definition
   * doesn't have the `renderable` capability.
   *
   */
  readonly seo: Maybe<MetaobjectSeo>;
  /** The type of the metaobject. Defines the namespace of its associated metafields. */
  readonly type: Scalars['String']['output'];
  /** The date and time when the metaobject was last updated. */
  readonly updatedAt: Scalars['DateTime']['output'];
};
// ...
export type Model3d = Media & Node & {
  readonly __typename: 'Model3d';
  /** A word or phrase to share the nature or contents of a media. */
  readonly alt: Maybe<Scalars['String']['output']>;
  /** A globally-unique ID. */
  readonly id: Scalars['ID']['output'];
  /** The media content type. */
  readonly mediaContentType: MediaContentType;
  /** The presentation for a media. */
  readonly presentation: Maybe<MediaPresentation>;
  /** The preview image for the media. */
  readonly previewImage: Maybe<Image>;
  /** The sources for a 3d model. */
  readonly sources: ReadonlyArray<Model3dSource>;
};
// ...
export type Video = Media & Node & {
  readonly __typename: 'Video';
  /** A word or phrase to share the nature or contents of a media. */
  readonly alt: Maybe<Scalars['String']['output']>;
  /** A globally-unique ID. */
  readonly id: Scalars['ID']['output'];
  /** The media content type. */
  readonly mediaContentType: MediaContentType;
  /** The presentation for a media. */
  readonly presentation: Maybe<MediaPresentation>;
  /** The preview image for the media. */
  readonly previewImage: Maybe<Image>;
  /** The sources for a video. */
  readonly sources: ReadonlyArray<VideoSource>;
};

/** Represents a source for a Shopify hosted video. */
export type VideoSource = {
  readonly __typename: 'VideoSource';
  /** The format of the video source. */
  readonly format: Scalars['String']['output'];
  /** The height of the video. */
  readonly height: Scalars['Int']['output'];
  /** The video MIME type. */
  readonly mimeType: Scalars['String']['output'];
  /** The URL of the video. */
  readonly url: Scalars['String']['output'];
  /** The width of the video. */
  readonly width: Scalars['Int']['output'];
};
// ...
export type CollectionsFragment = { readonly __typename: 'CollectionConnection', readonly edges: ReadonlyArray<{ readonly __typename: 'CollectionEdge', readonly node: { readonly __typename: 'Collection', readonly id: string, readonly title: string, readonly handle: string, readonly image: { readonly __typename: 'Image', readonly url: any } | null, readonly metafield: { readonly __typename: 'Metafield', readonly type: string, readonly reference: { readonly __typename: 'Collection' } | { readonly __typename: 'GenericFile', readonly url: any | null, readonly mimeType: string | null } | { readonly __typename: 'MediaImage' } | { readonly __typename: 'Metaobject' } | { readonly __typename: 'Model3d' } | { readonly __typename: 'Page' } | { readonly __typename: 'Product' } | { readonly __typename: 'ProductVariant' } | { readonly __typename: 'Video', readonly sources: ReadonlyArray<{ readonly __typename: 'VideoSource', readonly url: string, readonly mimeType: string, readonly format: string }>, readonly previewImage: { readonly __typename: 'Image', readonly url: any } | null } | null } | null, readonly tagsMetafield: { readonly __typename: 'Metafield', readonly type: string, readonly value: string } | null } }> };